requirements.txt
Flask
google-cloud-firestore
google-cloud-bigquery
gunicorn
Werkzeug
